interface/requirements.txt
PyPDF2==3.0.1
pyserial==3.5
Werkzeug==3.1.3
flask==3.1.0